styling asset cards and moving files around
