+++ web/src/components/assets/AssetCreationForm.module.css
@@ -1,6 +1,7 @@
 .container {
   width: 70%;
   margin: 0 auto;
+  margin-bottom: 2rem;
   padding: 2rem;
   border: 1px solid #e0e0e0;
   border-radius: 8px;
