+++ web/src/components/studies/ManageStudy.module.css
@@ -13,5 +13,6 @@
 }
 
 .completed-section {
+  width: 80%;
   margin-top: 2rem;
 }
