+++ web/src/components/studies/StudyAgreement.module.css
