+++ web/src/components/studies/StudyAgreement.tsx
