--- web/src/components/studies/study-steps/StudyAssets.module.css
@@ -1,26 +1,0 @@
-.study-assets-container {
-  width: 100%;
-  margin: 1rem 0;
-  padding: 1.5rem;
-  background: var(--background);
-  border-radius: 8px;
-  border: 1px solid #e5e7eb;
-}
-
-.error-message {
-  color: #d32f2f;
-  background-color: #fdeaea;
-  border: 1px solid #f5c6cb;
-  border-radius: 4px;
-  padding: 12px;
-  margin-bottom: 16px;
-}
-
-.no-assets-message {
-  color: #6c757d;
-  padding: 12px;
-  border: 1px solid #dee2e6;
-  border-radius: 4px;
-  background-color: #f8f9fa;
-  margin-bottom: 16px;
-}
